mt_tower/requirements.txt
aioboto3
discord.py
